website/requirements/requirements.txt
# THIS IS THE 'REAL' ONE !!

# pinax related
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.3.1
-e git://github.com/pinax/pinax.git@c90f20ef9e5ed3a890cb2524634075477b7d35fc#egg=Pinax-dev

PIL==1.1.7

django-debug-toolbar==0.8.4
django-staticfiles==1.0.1
django-compressor==0.7.1

django-mailer==0.2a1.dev3
django-email-confirmation==0.2.dev6
django-timezones==0.2.dev2
pytz==2011h
django-openid==0.2.3.dev1
python-openid==2.2.5
django-uni-form==0.7.0

kaleo==0.2
django_polymorphic

django-extensions
vobject
django-tinymce
chardet

# hm what version?
easy-thumbnails==1.0-alpha-18

django-announcements==0.1.2
django-pagination==1.0.7
django-notification==0.2a1.dev4
idios==0.2a1.dev2

# other
wsgiref==0.1.2

# cms
# MySQL-python==1.2.3
South==0.7.3
django-classy-tags==0.3.4.1
django-cms==2.2
django-mptt==0.4.2
django-flashpolicies==1.4.1


# additional
django-dajaxice==0.2
django-filer==0.8.3
cmsplugin-filer==0.8.0
html5lib==0.90
flup

gunicorn
gunicorn-console
